Убрал функцию создающую данные в модуль
